# web-api/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Webcomics Platform API
  version: 0.1.0
  description: |
    HTTP API for the webcomics generation platform. Two route groups:
      * `/users/*`  — user CRUD (boilerplate carry-over)
      * `/api/*`    — pipeline runs, templates, assets
servers:
  - url: http://localhost:8080
paths:
  /users:
    get:
      operationId: listUsers
      summary: List users
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/UserView' }
        default: { $ref: '#/components/responses/Error' }
    post:
      operationId: registerUser
      summary: Register a user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterUserRequest' }
      responses:
        '201':
          description: created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IDResponse' }
        default: { $ref: '#/components/responses/Error' }
  /users/{id}:
    get:
      operationId: getUser
      summary: Get a user
      parameters:
        - $ref: '#/components/parameters/PathID'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserView' }
        default: { $ref: '#/components/responses/Error' }
  /users/{id}/activate:
    post:
      operationId: activateUser
      summary: Activate a pending user
      parameters:
        - $ref: '#/components/parameters/PathID'
      responses:
        '204':
          description: no content
        default: { $ref: '#/components/responses/Error' }
  /api/runs:
    get:
      operationId: listRuns
      summary: List pipeline runs
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 50, maximum: 200 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
        - in: query
          name: status
          description: Comma-separated status filter (e.g. "running,completed").
          schema: { type: string }
        - in: query
          name: q
          description: Case-insensitive substring match on the prompt.
          schema: { type: string }
        - in: query
          name: project_id
          description: Filter by project. Omit to include runs without a project too.
          schema: { type: string }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/RunSummary' }
        default: { $ref: '#/components/responses/Error' }
    post:
      operationId: createRun
      summary: Create and start a pipeline run
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRunRequest' }
      responses:
        '201':
          description: created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IDResponse' }
        default: { $ref: '#/components/responses/Error' }
...
components:
  parameters:
    PathID:
      in: path
      name: id
      required: true
      schema: { type: string }
  responses:
    Error:
      description: error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
    IDResponse:
      type: object
      required: [id]
      properties:
        id: { type: string }
    RegisterUserRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    UserView:
      type: object
      required: [id, email, status, created_at]
      properties:
        id: { type: string }
        email: { type: string }
        status: { type: string }
        created_at: { type: string, format: date-time }
...
    CreateRunRequest:
      type: object
      required: [prompt, template_id]
      properties:
        prompt: { type: string }
        template_id: { type: string }
        format_id:
          type: string
          description: Format preset id (e.g. "manga", "webtoon"). Overrides project default format.
        project_id:
          type: string
          description: Optional. Attaches run to a project; characters/environments below select which ones influence this run.
        character_ids:
          type: array
          description: Subset of the project's characters to apply.
          items: { type: string }
        environment_ids:
          type: array
          description: Subset of the project's environments to apply.
          items: { type: string }
        use_plot:
          type: boolean
          description: Include the project's plot premise+beats in script payload.
        overrides: { $ref: '#/components/schemas/RunOverrides' }
    RunOverrides:
      type: object
      description: |
        Per-run tweaks applied on top of the chosen template. All fields are
        optional. `steps` replaces the whole snapshot when present; the other
        fields are convenience tweaks merged into the template's steps.
      properties:
        panel_count:
          type: integer
          minimum: 1
          maximum: 20
          description: Number of panels (overrides script step's params.panel_count).
        target_duration_ms:
          type: integer
          minimum: 1000
          maximum: 120000
          description: Total final video length, distributed across panels.
        enable_audio:
          type: boolean
          description: Insert an audio step between image and assemble.
        system_prompt:
          type: string
          description: Override script step's system_prompt.
        script_model:
          type: string
          description: Override script step's model.
        image_model:
          type: string
          description: Override image step's model.
        style_reference:
          type: string
          enum: [none, anchor, previous]
          description: |
            Image-step style consistency mode. none = parallel (current default);
            anchor = panel 0 referenced by all later panels; previous = cumulative
            (panel N references 0..N-1). Sequential modes require a ref-capable model.
        auto_assemble:
          type: boolean
          description: |
            When false the run pauses in awaiting_action state before the
            assemble step instead of auto-rendering. The user then calls
            POST /api/runs/{id}/assemble to render. Default true.
        audio:
          type: object
          description: Audio step overrides (voice + model + speed). Applied when enable_audio=true.
          properties:
            voice_id: { type: string }
            model: { type: string }
            speed: { type: number, format: double, minimum: 0.5, maximum: 2.0 }
        upload:
          type: object
          description: Auto-upload settings. Enables caption+upload steps after assemble.
          properties:
            enabled: { type: boolean }
            provider: { type: string, description: youtube_selenium | twitter_selenium }
            social_account_ids:
              type: array
              items: { type: string }
            scheduled_at:
              type: string
              format: date-time
              description: RFC3339; if in the future, worker waits.
            caption_model: { type: string }
            caption_override:
              type: string
              description: Skip LLM, use this caption for all platforms.
            platforms:
              type: array
              items: { type: string }
        assemble:
          type: object
          description: Assemble step overrides (fps + resolution + codec).
          properties:
            fps: { type: integer, minimum: 1, maximum: 120 }
            width: { type: integer, minimum: 120 }
            height: { type: integer, minimum: 120 }
            codec: { type: string, enum: [h264, h265] }
        steps:
          type: array
          description: Full step override (replaces template snapshot when non-empty).
          items: { $ref: '#/components/schemas/StepConfig' }
...
    StepConfig:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [script, image, audio, music, assemble, upload]
        system_prompt: { type: string }
        model: { type: string }
        provider: { type: string }
        params:
          type: object
          additionalProperties: true
...
    RunSummary:
      type: object
      required: [id, template_id, prompt, status, total_cost_usd, created_at]
      properties:
        id: { type: string }
        template_id: { type: string }
        prompt: { type: string }
        status: { type: string }
        total_cost_usd: { type: number, format: double }
        created_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time, nullable: true }
        video_asset_id:
          type: string
          description: ID of the final assembled video asset, if one was produced.
...